2. promote controller
