+++ pkg/controller/direct/apigateway/api_controller.go
@@ -16,7 +16,7 @@
 // proto.service: google.cloud.apigateway.v1.ApiGatewayService
 // proto.message: google.cloud.apigateway.v1.Api
 // crd.type: APIGatewayAPI
-// crd.version: v1alpha1
+// crd.version: v1beta1
 
 package apigateway
 
@@ -33,7 +33,7 @@ import (
 	"k8s.io/klog/v2"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	krm "github.com/GoogleCloudPlatform/k8s-config-connector/apis/apigateway/v1alpha1"
+	krm "github.com/GoogleCloudPlatform/k8s-config-connector/apis/apigateway/v1beta1"
 	refs "github.com/GoogleCloudPlatform/k8s-config-connector/apis/refs/v1beta1"
 	"github.com/GoogleCloudPlatform/k8s-config-connector/pkg/config"
 	"github.com/GoogleCloudPlatform/k8s-config-connector/pkg/controller/direct"
